Add linting and type checking
Added a GitHub workflow to run pre-commit, pylint, and pyright.

pylint and pyright are not included as pre-commit checks, because they
need to analyze the project as a whole, not just the changed files, and
they do not work correctly in an isolated environment without having
access to dependencies.

Adjusted the pylint configuration to ignore diagnostics that are either
not useful or duplicate those reported by pyright.

Fixed all type errors reported by pyright.

Added VS Code extension recommendations for all the formatting and
linting tools, and enabled type checking in VS Code.

Fixes #3